fix: cleanups, float RRE, refactor
